requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cssselect2==0.7.0
Django==4.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crontab==0.7.1
django-el-pagination==4.0.0
django-js-asset==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
html5lib==1.1
httplib2==0.22.0
idna==3.6
lxml==5.1.0
msgpack==1.0.8
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.2.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.2
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
xhtml2pdf==0.2.15
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
